+++ fastapi_forge/frontend/constants.py
