+++ fastapi_forge/frontend/modals/__init__.py
@@ -1,0 +1,5 @@
+from fastapi_forge.frontend.modals.field_modal import AddFieldModal, UpdateFieldModal
+from fastapi_forge.frontend.modals.relation_modal import (
+    AddRelationModal,
+    UpdateRelationModal,
+)
