+++ fastapi_forge/frontend/notifications.py
@@ -1,5 +1,4 @@
-from pydantic import BaseModel, ValidationError
-from fastapi_forge import dtos as m
+from pydantic import ValidationError
 from nicegui import ui
