+++ pyproject.toml
@@ -20,7 +20,7 @@ Repository = "https://github.com/mslaursen/fastapi-forge.git"
 fastapi-forge = "fastapi_forge.__main__:main"
 
 [tool.ruff]
-exclude = ["template"]
+exclude = ["template", "__init__.py"]
 line-length = 88
 indent-width = 4
